// apps/api/src/ships/sheet.types.ts
export interface ShipSheet {
  pot: number;
  pdf: number;
  bli: number;
  esc: number;
  sen: number;
  crg: number;
  min: number;
  hp: number;
  mass: number;
  energyCont: number;
  energyCombat: number;
  batCharge: number;
  batOutput: number;
  batInput: number;
  fuelCap: number;
  fuelUse: number;
  structureUsed: number;
  structureBudget: number;
  autonomy: number;
  mob: number;
  condition: number;
}